Refactor IPAM driver
Abstract out IPAM function into it's own class. At the same time,
relocate attributes from ObjectManager to the new InfobloxDriver
class.

Also, rename the executable to be consistent with the naming of
the container.
